Working version of interpolate.py
Bug fix included adding eps to bounds check in get_lm() since it was failing on
points on the cell face.

Added CellTree and CellTree_interpolator as classes to do interpolation, separating
the creation of the tree, the location of the points, and interpolating the field.

Added test functions to test cell_tree2d (to see if the point is in the polygon,
within a specified tolerance, eps), and a test of interpolation using an analytical
function for z.